:truck: Moved network state to `.shared.SharedStates`
