+++ base/src/main/java/com/arnyminerz/escalaralcoiaicomtat/fragment/model/NetworkChangeListenerFragment.kt
@@ -1,8 +1,8 @@
 package com.arnyminerz.escalaralcoiaicomtat.fragment.model
 
 import androidx.fragment.app.Fragment
-import com.arnyminerz.escalaralcoiaicomtat.appNetworkProvider
 import com.arnyminerz.escalaralcoiaicomtat.network.base.ConnectivityProvider
+import com.arnyminerz.escalaralcoiaicomtat.shared.appNetworkProvider
 
 abstract class NetworkChangeListenerFragment : Fragment(),
     ConnectivityProvider.ConnectivityStateListener {
